move duplicated code of chain extension to runtime-common
