+++ runtime/common/Cargo.toml
@@ -12,9 +12,12 @@ targets = ["x86_64-unknown-linux-gnu"]
 
 [dependencies]
 # Substrate
+parity-scale-codec = {version = "3.1.5", default-features = false, features = ["derive"]}
+scale-info = {version = "2.1.1", default-features = false, features = ["derive"]}
 sp-runtime = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.37" }
 sp-consensus-aura = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.37" }
 sp-core = { git = "https://github.com/paritytech/substrate", default-features = false, branch = "polkadot-v0.9.37" }
+spacewalk-primitives = { git = "https://github.com/pendulum-chain/spacewalk", default-features = false, branch = "dev/milestone-3" }
 
 [features]
 default = [
@@ -25,6 +28,9 @@ std = [
 	"sp-consensus-aura/std",
 	"sp-runtime/std",
 	"sp-core/std",
+	"spacewalk-primitives/std",
+	"parity-scale-codec/std",
+	"scale-info/std"
 ]
 
 runtime-benchmarks = [
